+++ tests/unit/utils/type-safe-get-test.ts
@@ -1,4 +1,4 @@
-import typeSafeGet, { overrideGet } from 'ember-typescript-utils/type-safe-get';
+import typeSafeGet from 'ember-typescript-utils/type-safe-get';
 import { module, test } from 'qunit';
 
 module('Unit | Utility | type-safe-get', function(_hooks) {
@@ -14,19 +14,6 @@ module('Unit | Utility | type-safe-get', function(_hooks) {
     assert.deepEqual(typeSafeGet(test, 'a', 'b', 'c', 'd', 'e', 'f'), test.a.b.c.d.e.f);
     assert.deepEqual(typeSafeGet(test, 'a', 'b', 'c', 'd', 'e', 'f', 'g'), test.a.b.c.d.e.f.g);
     assert.deepEqual(typeSafeGet(test, 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h'), test.a.b.c.d.e.f.g.h);
-  });
-
-  test('overrideGet also works', function(assert) {
-    const test = {a: {b: {c: {d: {e: {f: {g: {h: 'made it!'}}}}}}}};
-    
-    assert.deepEqual(overrideGet(test, 'a'), test.a);
-    assert.deepEqual(overrideGet(test, 'a', 'b'), test.a.b);
-    assert.deepEqual(overrideGet(test, 'a', 'b', 'c'), test.a.b.c);
-    assert.deepEqual(overrideGet(test, 'a', 'b', 'c', 'd'), test.a.b.c.d);
-    assert.deepEqual(overrideGet(test, 'a', 'b', 'c', 'd', 'e'), test.a.b.c.d.e);
-    assert.deepEqual(overrideGet(test, 'a', 'b', 'c', 'd', 'e', 'f'), test.a.b.c.d.e.f);
-    assert.deepEqual(overrideGet(test, 'a', 'b', 'c', 'd', 'e', 'f', 'g'), test.a.b.c.d.e.f.g);
-    assert.deepEqual(overrideGet(test, 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h'), test.a.b.c.d.e.f.g.h);
   });
 
   test('it returns undefined if the initial object is undefined', function(assert) {