+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Moved `overrideGet` from additional export of `ember-typescript-utils/type-safe-get` to default export of `ember-typescript-utils/override-get`
+
 ### Added
 - YUI docs
 - This changelog